@@ -11,62 +11,65 @@
         "prepreview": "pnpm run build",
         "preview": "astro preview",
         "astro": "astro",
-        "create:post": "node scripts/createPost.mjs",
+        "post": "node scripts/createPost.mjs",
         "create:project": "node scripts/createProject.mjs",
-        "check:lint": "eslint --ignore-path .gitignore .",
-        "lint": "eslint --ignore-path .gitignore --fix .",
+        "check:lint": "eslint .",
+        "lint": "eslint --fix .",
         "check:format": "prettier --list-different  \"**/*.{js,jsx,ts,tsx,md,mdx,astro}\"",
         "format": "prettier --write  \"**/*.{js,jsx,ts,tsx,md,mdx,astro}\"",
         "validate": "npm-run-all --parallel check:*",
-        "postbuild": "pagefind"
+        "postbuild": "pagefind",
+        "prepare": "husky"
     },
     "dependencies": {
-        "@astrojs/check": "^0.9.3",
-        "@astrojs/react": "^3.6.2",
-        "@astrojs/rss": "^4.0.7",
-        "@astrojs/sitemap": "^3.1.6",
-        "@astrojs/tailwind": "^5.1.1",
-        "@phosphor-icons/react": "^2.1.7",
-        "@playform/inline": "^0.0.3",
-        "@types/react": "^18.3.7",
-        "@types/react-dom": "^18.3.0",
-        "astro": "^4.15.8",
+        "@astrojs/check": "^0.9.4",
+        "@astrojs/react": "^4.3.0",
+        "@astrojs/rss": "^4.0.12",
+        "@astrojs/sitemap": "^3.5.0",
+        "@phosphor-icons/react": "^2.1.10",
+        "@playform/inline": "^0.1.2",
+        "@tailwindcss/vite": "^4.1.12",
+        "@types/react": "^19.1.10",
+        "@types/react-dom": "^19.1.7",
+        "astro": "^5.13.2",
         "clsx": "^2.1.1",
         "mdast-util-to-string": "^4.0.0",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
+        "react": "^19.1.1",
+        "react-dom": "^19.1.1",
         "reading-time": "^1.5.0",
-        "satori": "^0.10.14",
+        "satori": "^0.16.2",
         "satori-html": "^0.3.2",
-        "sharp": "^0.33.5",
-        "tailwindcss": "^3.4.12",
-        "tm-themes": "^1.8.2",
-        "typescript": "^5.6.2"
+        "sharp": "^0.34.3",
+        "tailwindcss": "^4.1.12",
+        "tm-themes": "^1.10.7"
     },
     "devDependencies": {
-        "@pagefind/default-ui": "^1.1.1",
+        "@eslint/compat": "^1.3.2",
+        "@eslint/js": "^9.33.0",
+        "@pagefind/default-ui": "^1.3.0",
         "@resvg/resvg-js": "^2.6.2",
-        "@tailwindcss/typography": "^0.5.15",
-        "@types/inquirer": "^9.0.7",
+        "@tailwindcss/typography": "^0.5.16",
+        "@types/inquirer": "^9.0.9",
         "@types/json-to-pretty-yaml": "^1.2.1",
-        "@typescript-eslint/eslint-plugin": "^7.18.0",
-        "@typescript-eslint/parser": "^7.18.0",
-        "chalk": "^5.3.0",
-        "date-fns": "^3.6.0",
-        "eslint": "^8.57.1",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-astro": "^0.31.4",
-        "eslint-plugin-jsx-a11y": "^6.10.0",
-        "eslint-plugin-simple-import-sort": "^12.1.1",
-        "inquirer": "^9.3.6",
+        "chalk": "^5.6.0",
+        "date-fns": "^4.1.0",
+        "eslint": "^9.33.0",
+        "eslint-config-prettier": "^10.1.8",
+        "eslint-plugin-astro": "^1.3.1",
+        "globals": "^16.3.0",
+        "husky": "^9.1.7",
+        "inquirer": "^12.9.3",
         "json-to-pretty-yaml": "^1.2.2",
+        "lint-staged": "^16.1.5",
         "npm-run-all": "^4.1.5",
-        "pagefind": "^1.1.1",
-        "prettier": "^3.3.3",
-        "prettier-plugin-astro": "^0.13.0",
-        "prettier-plugin-tailwindcss": "^0.5.14",
-        "tailwind-merge": "^2.5.2",
+        "pagefind": "^1.3.0",
+        "prettier": "3.6.2",
+        "prettier-plugin-astro": "^0.14.1",
+        "prettier-plugin-tailwindcss": "^0.6.14",
+        "tailwind-merge": "^3.3.1",
         "title-case": "^4.3.2",
+        "typescript": "^5.9.2",
+        "typescript-eslint": "^8.40.0",
         "yamljs": "^0.3.0"
     }
 }
