@@ -1,4 +1,4 @@
-module.exports = {
+export default {
     '*.+(js)': ['eslint'],
     '*.+(js|json)': ['prettier --write'],
 };
