@@ -1,7 +1,7 @@
 import react from '@astrojs/react';
 import sitemap from '@astrojs/sitemap';
-import tailwind from '@astrojs/tailwind';
 import playformInline from '@playform/inline';
+import tailwindcss from '@tailwindcss/vite';
 import { defineConfig } from 'astro/config';
 
 import { remarkPostMeta } from './src/utils/remark-post-meta.mjs';
@@ -10,19 +10,15 @@ import { remarkPostMeta } from './src/utils/remark-post-meta.mjs';
 export default defineConfig({
     site: 'https://shahid.pro/',
     base: '/',
-    integrations: [
-        tailwind({
-            applyBaseStyles: false,
-        }),
-        react(),
-        sitemap(),
-        playformInline(),
-    ],
+    vite: {
+        plugins: [tailwindcss()],
+    },
+    integrations: [react(), sitemap(), playformInline()],
     markdown: {
         remarkPlugins: [remarkPostMeta],
         shikiConfig: {
             themes: {
-                light: 'snazzy-light',
+                light: 'catppuccin-latte',
                 dark: 'tokyo-night',
             },
             langs: [],
