@@ -155,7 +155,7 @@ export async function GET(context: APIContext) {
         : markup(title, publishedAt, minutesRead);
     const svg = await satori(html as ReactNode, ogOptions);
     const png = new Resvg(svg).render().asPng();
-    return new Response(png, {
+    return new Response(new Uint8Array(png), {
         headers: {
             'Cache-Control': 'public, max-age=31536000, immutable',
             'Content-Type': 'image/png',
